step3: implements DiscountCondition  다형성패턴, 변경 보호패턴
