+++ src/chapter5/Movie.ts
@@ -51,20 +51,7 @@ export class Movie {
   }
 
   private isDiscountable(screening: Screening): boolean {
-    return (
-      this.checkPeriodConditions(screening) ||
-      this.checkSequenceConditions(screening)
-    );
-  }
-
-  private checkPeriodConditions(screening: Screening): boolean {
-    return this.periodConditions.some((condition) =>
-      condition.isSatisfiedBy(screening)
-    );
-  }
-
-  private checkSequenceConditions(screening: Screening): boolean {
-    return this.sequenceConditions.some((condition) =>
+    return this.discountConditions.some((condition) =>
       condition.isSatisfiedBy(screening)
     );
   }
